+++ chainweb.cabal
@@ -143,6 +143,7 @@ library
         , Chainweb.Mempool.RestAPI.Server
         , Chainweb.MerkleLogHash
         , Chainweb.MerkleUniverse
+        , Chainweb.Miner
         , Chainweb.Miner.Config
         , Chainweb.Miner.POW
         , Chainweb.Miner.Test
@@ -226,7 +227,6 @@ library
         , Chainweb.Pact.Backend.Types
         , Chainweb.Pact.Backend.Utils
         , Chainweb.Pact.BloomCache
-        , Chainweb.Pact.Miner
         , Chainweb.Pact.PactService
         , Chainweb.Pact.RestAPI
         , Chainweb.Pact.RestAPI.Server
