+++ test/Chainweb/Test/Pact/PactExec.hs
@@ -33,8 +33,8 @@ import Test.Tasty.HUnit
 
 import Chainweb.BlockHash
 import Chainweb.BlockHeader.Genesis
+import Chainweb.Miner
 import Chainweb.Pact.Backend.Types
-import Chainweb.Pact.Miner
 import Chainweb.Pact.PactService
 import Chainweb.Pact.Types
 import Chainweb.Payload.PayloadStore.InMemory
